Factor out common API
